fixed classroom

+++ software/epe/classroom/views.py
@@ -7,13 +7,13 @@ def list(request):
     if not request.user.is_authenticated:
         return redirect('%s?next=%s' % (settings.LOGIN_REDIRECT_URL, request.path))
     classrooms = Classroom.objects.all()
-    return render(request, 'list.html', {'classrooms': classrooms, 'request': request})
+    return render(request, '../templates/classroom_list.html', {'classrooms': classrooms, 'request': request})
 
 def detail(request, id):
     if not request.user.is_authenticated:
         return redirect('%s?next=%s' % (settings.LOGIN_REDIRECT_URL, request.path))
     classroom = Classroom.objects.get(code=id)
-    return render(request, 'detail.html', {'classroom': classroom})
+    return render(request, '../templates/classroom_detail.html', {'classroom': classroom})
 
 def create(request):
     if not request.user.is_authenticated:
@@ -25,39 +25,35 @@ def create(request):
         classroom.code = code
         classroom.name = name
         if not code:
-            return render(request, 'create.html', {'classroom': classroom, 'error': "Código del aula no introducido"})
+            return render(request, '../templates/classroom_create.html', {'classroom': classroom, 'error': "Código del aula no introducido"})
         if not name:
-            return render(request, 'create.html', {'classroom': classroom, 'error': "Nombre del aula no introducido"})
+            return render(request, '../templates/classroom_create.html', {'classroom': classroom, 'error': "Nombre del aula no introducido"})
         if Classroom.objects.filter(code=code).exists():
-            return render(request, 'create.html', {'classroom': classroom, 'error': "Código del aula ya existe"})
+            return render(request, '../templates/classroom_create.html', {'classroom': classroom, 'error': "Código del aula ya existe"})
         classroom.save()
         request.session['info'] = "Aula agregada"
-        return render(request, "detail.html", {'classroom': classroom, 'info': "Aula agregada"})
-    return render(request, 'create.html')
+        return render(request, "../templates/classroom_detail.html", {'classroom': classroom, 'info': "Aula agregada"})
+    return render(request, '../templates/classroom_create.html')
 
 def update(request, id):
     if not request.user.is_authenticated:
         return redirect('%s?next=%s' % (settings.LOGIN_REDIRECT_URL, request.path))
     classroom = Classroom.objects.get(code=id)
     if request.POST:
-        code = request.POST['code']
         name = request.POST['name']
-        if not code:
-            return render(request, 'update.html', {'classroom': classroom, 'error': "Código del aula no introducido"})
         if not name:
-            return render(request, 'update.html', {'classroom': classroom, 'error': "Nombre del aula no introducido"})
-        if classroom.code != code and Classroom.objects.filter(code=code).exists():
-            return render(request, 'update.html', {'classroom': classroom, 'error': "Código del aula ya existe"})
-        classroom.code = code
+            return render(request, '../templates/classroom_update.html', {'classroom': classroom, 'error': "Nombre del aula no introducido"})
         classroom.name = name
         classroom.save()
-        return render(request, 'detail.html', {'classroom': classroom, 'info': "Aula "+classroom.name+" actualizada correctamente"})
-    return render(request, 'update.html', {'classroom': classroom})
+        return render(request, '../templates/classroom_detail.html', {'classroom': classroom, 'info': "Aula " + classroom.name + " actualizada correctamente"})
+    return render(request, '../templates/classroom_update.html', {'classroom': classroom})
 
 def delete(request, id):
     if not request.user.is_authenticated:
         return redirect('%s?next=%s' % (settings.LOGIN_REDIRECT_URL, request.path))
     classroom = Classroom.objects.get(code=id)
+    if classroom.devices.all().count() > 0:
+        return render(request, '../templates/classroom_detail.html', {'classroom': classroom, 'error': "El aula " + classroom.name + " tiene dispositivos asociados"})
     classroom.delete()
     classrooms = Classroom.objects.all()
-    return render(request, 'list.html', {'classrooms': classrooms, 'info': "Aula "+classroom.name+" borrada"})
+    return render(request, '../templates/classroom_list.html', {'classrooms': classrooms, 'info': "Aula " + classroom.name + " borrada"})
